backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pillow>=10.4.0
transformers>=4.44.0
torch>=2.0.0
faiss-cpu>=1.8.0
